First commit; max leaf and vendor ID test
This is a small project to test the reading of information from CPUID,
and possibly other sources.  I may integrate TSC-related content from
"optiflop" into this library someday.

Also, it's not a library yet, just object code + a small test program.

There are *many* other similar projects, but this is a slightly more
object-oriented effort to see if a more dynamic and interactive "Cpuid"
object adds any value.